+++ internal/app/models/order.go
@@ -16,7 +16,7 @@ type OrderID = string
 type OrderItem struct {
 	OrderID    OrderID     `json:"number"`
 	Status     OrderStatus `json:"status"`
-	Accrual    *uint64     `json:"accrual,omitempty"`
+	Accrual    *float32    `json:"accrual,omitempty"`
 	UploadTime time.Time   `json:"uploaded_at"`
 }
 type Orders []OrderItem
@@ -46,6 +46,6 @@ type AccrualOrderItem struct {
 	OrderID OrderID            `json:"order"`
 	UserID  UserID             `json:"-"`
 	Status  AccrualOrderStatus `json:"status"`
-	Accrual *uint64            `json:"accrual,omitempty"`
+	Accrual *float32           `json:"accrual,omitempty"`
 	Error   error              `json:"-"`
 }
